requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==4.2.13
django-autoslug==1.9.9
django-crispy-forms==2.2
html5lib==1.1
idna==3.7
lxml==5.2.2
oauthlib==3.2.2
oscrypto==1.3.0
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python3-openid==3.2.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
